adding animations and testing them
adding animations and testing them - some issues are popping up in the animations, the dash isn't always triggering, besides that I need to just work on the dahs animation, trying to figure out the way way to go about it. Think I will keep it simpler then current version which is way too slow.
Fall back is instead turning it into a roll