tilesets, fog again and prefabs
Added the tileset and updated it until the two levels could be filled out using it.
updated duplicate versions of the combat and movement levels with the tileset, I duplicated because I didnt want to destroy anything.
continued on fog (i need to end this...)
updated the enemy prefabs with our assets